Revert "updated Tracker function to pass goal to App"
