+++ package.json
@@ -3,8 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "firebase": "^6.6.0",
-    "node-sass": "^4.12.0",
+    "axios": "^0.19.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-scripts": "3.1.1"
@@ -31,7 +30,6 @@
     ]
   },
   "devDependencies": {
-    "axios": "^0.19.0",
     "qs": "^6.8.0"
   }
 }
