--- src/components/Header.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-
-class Header extends Component {
-
-  render() {
-    return (
-      <div>Header</div>
-    )
-  }
-}
-
-export default Header
