--- src/partials/_displayBooks.scss
@@ -1,10 +1,0 @@
-.displayBooksContainer {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
-.bookImages {
-    margin: 4px;
-}
