--- src/partials/_global.scss
@@ -1,8 +1,0 @@
-// colors
-$brightBlue: #5C67FF;
-$lightBlue: #C8E0F8;
-$darkBlue: #1B2456;
-$blush: #E5C1BD;
-$cranberry: #E25050;
-
-
